@@ -17,7 +17,7 @@
     "pg": "^8.3.0",
     "reflect-metadata": "^0.1.10",
     "tsyringe": "^4.2.0",
-    "typeorm": "0.2.24"
+    "typeorm": "^0.2.32"
   },
   "devDependencies": {
     "@types/cors": "^2.8.6",
@@ -38,8 +38,8 @@
     "supertest": "^4.0.2",
     "ts-jest": "^25.5.1",
     "ts-node": "3.3.0",
-    "ts-node-dev": "^1.0.0-pre.44",
+    "ts-node-dev": "^1.1.6",
     "tsconfig-paths": "^3.9.0",
-    "typescript": "^3.8.3"
+    "typescript": "^4.2.0"
   }
 }
